импортирую постоянные переменные из util.js
