+++ js/util.js
@@ -1,0 +1,25 @@
+const MIN_LIKES = 15;
+const MAX_LIKES = 200;
+const CANVAS_SIZE = 25;
+
+const COMMENTS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const RANDOM_NAMES = [
+  'Хуан',
+  'Себастьян',
+  'Мария',
+  'Кристоф',
+  'Виктор',
+  'Юлия',
+  'Люпита',
+  'Вашингтон',
+];
+export {MIN_LIKES, MAX_LIKES, CANVAS_SIZE, COMMENTS, RANDOM_NAMES};
+//
